[addync_sync_to_suggestions] Adding sync to Django/Python suggestions
